+++ tests/TestCase.php
@@ -126,8 +126,8 @@ class TestCase extends Orchestra
      */
     public function configureDatabase($app)
     {
-        DB::statement('DROP TABLE IF EXISTS test_users CASCADE;');
         DB::statement('DROP TABLE IF EXISTS test_permissions CASCADE;');
+        DB::statement('DROP TABLE IF EXISTS test_users CASCADE;');
         DB::statement('DROP TABLE IF EXISTS test_groups CASCADE;');
         DB::statement('DROP TABLE IF EXISTS test_user_has_permissions CASCADE;');
         DB::statement('DROP TABLE IF EXISTS test_user_has_groups CASCADE;');
@@ -169,6 +169,11 @@ class TestCase extends Orchestra
         /*
          * Create the tables on the database
          */
+        (new \CreatePermissionsTable())->down();
+        (new \CreateGroupsTable())->down();
+        (new \CreateGroupHasPermissionsTable())->down();
+        (new \CreateUserHasPermissionsTable())->down();
+        (new \CreateUserHasGroupsTable())->down();
         (new \CreatePermissionsTable())->up();
         (new \CreateGroupsTable())->up();
         (new \CreateGroupHasPermissionsTable())->up();
@@ -240,6 +245,16 @@ class TestCase extends Orchestra
             'name' => 'Edit news',
             'slug' => 'edit-news',
             'description' => 'This permission allows you to edit the news page',
+        ]);
+        Permission::create([
+            'name' => 'Test hierarchical permissions',
+            'slug' => 'admin.auth',
+            'description' => 'This is a hierarchical permission test',
+        ]);
+        Permission::create([
+            'name' => 'Test hierarchical permissions 1',
+            'slug' => 'admin.auth.users',
+            'description' => 'This is a hierarchical permission test',
         ]);
     }
 }